Added apply mixin trait for use on options valuetypes.
Depending on how the option values are implemented you may or may
not want an apply method on the companion object.

The "of" method is always there by default in the Options trait, and
performs the same purpose, but without Scala's syntactic sugar that
is available for apply methods.

If the options are implemented using a case class, then you already
have an apply method.

However if the options are implemented using case objects inheriting
a common (normally sealed) trait, then having an apply method is useful.

The Apply trait (and subtraits StringApply and IntApply) allows an apply
method to be easily mixed into Options companion objects.