Fixed deserialization bug.
When deserializing a value with a limited range (i.e. where all
values of the underlying primitive type weren't valid, such as an
enum), the exception produced by the class itself would be thrown, most
commonly an IllegalArgumentException.

This conversion from 'primitive' type to the valuetype is now wrapped
in a try block and returned as a JsError instead of an exception being
thrown.